LayoutTests/http/tests/security/contentSecurityPolicy/report-uri-from-child-frame-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because of Content-Security-Policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
CSP report received:
CONTENT_TYPE: application/json
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/resources/generate-csp-report.html
REQUEST_METHOD: POST
=== POST DATA ===
{"csp-report":{"document-uri":"http://127.0.0.1:8000/security/contentSecurityPolicy/resources/generate-csp-report.html","violated-directive":"script-src 'self'"}}
